[frontend] Split user_helper out of webui_helper
the WebuiHelper grew too large, so separating functions
that render users
